add: refactor out to postulate
