+++ src/composer/mod.rs
@@ -2,3 +2,4 @@
 pub mod lambda_to_types;
 pub mod structures;
 pub mod agdaify;
+pub mod postulate;
